Expose ManualInputManager for TestCases to use
