+++ osu.Framework.Tests/Visual/TestCaseNestedMenus.cs
@@ -9,13 +9,10 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input;
-using osu.Framework.Input.Handlers;
-using osu.Framework.Platform;
 using osu.Framework.Testing;
+using osu.Framework.Testing.Input;
 using OpenTK;
 using OpenTK.Input;
-using MouseState = osu.Framework.Input.MouseState;
 
 namespace osu.Framework.Tests.Visual
 {
@@ -414,63 +411,6 @@ namespace osu.Framework.Tests.Visual
 
             item.Items = subItems;
             return item;
-        }
-
-        private class ManualInputManager : PassThroughInputManager
-        {
-            private readonly ManualInputHandler handler;
-
-            public ManualInputManager()
-            {
-                UseParentState = true;
-                AddHandler(handler = new ManualInputHandler());
-            }
-
-            public void MoveMouseTo(Drawable drawable)
-            {
-                UseParentState = false;
-                MoveMouseTo(drawable.ToScreenSpace(drawable.LayoutRectangle.Centre));
-            }
-
-            public void MoveMouseTo(Vector2 position)
-            {
-                UseParentState = false;
-                handler.MoveMouseTo(position);
-            }
-
-            public void Click(MouseButton button)
-            {
-                UseParentState = false;
-                handler.Click(button);
-            }
-        }
-
-        private class ManualInputHandler : InputHandler
-        {
-            private Vector2 lastMousePosition;
-
-            public void MoveMouseTo(Vector2 position)
-            {
-                PendingStates.Enqueue(new InputState { Mouse = new MouseState { Position = position } });
-                lastMousePosition = position;
-            }
-
-            public void Click(MouseButton button)
-            {
-                var mouseState = new MouseState { Position = lastMousePosition };
-                mouseState.SetPressed(button, true);
-
-                PendingStates.Enqueue(new InputState { Mouse = mouseState });
-
-                mouseState = (MouseState)mouseState.Clone();
-                mouseState.SetPressed(button, false);
-
-                PendingStates.Enqueue(new InputState { Mouse = mouseState });
-            }
-
-            public override bool Initialize(GameHost host) => true;
-            public override bool IsActive => true;
-            public override int Priority => 0;
         }
 
         /// <summary>
